addons/imora_project_task_ext/models/project_project.py
from odoo import api, fields, models


class ProjectProject(models.Model):
    _inherit = "project.project"

    code = fields.Char(
        string="Project Code",
    )
    task_sequence = fields.Integer(
        string="Task Sequence Counter",
        default=0,
    )
